scanner-service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
psutil==5.9.5
requests==2.31.0
scikit-learn==1.3.0
pydantic==2.6.1
python-dotenv==1.0.1
aiofiles==23.2.1
joblib==1.3.2
numpy==1.24.4
asyncio==3.4.3
httpx==0.26.0
clamd==1.0.2
python-dotenv==1.0.1
